Fixed Socket Transform (Right Hand)
Finally realized that you can use sockets to preview meshes in the skeleton window. Used this to cross reference my character with the example character from Mixamo, and now all the animations work! Still getting issues with an occasionally floating left hand but those issues seem to be present on Mixamo too so I'm not too fussed about them at the moment.