Attack Animation Swings once + Dodge Roll Junk
Forgot to commit for a while so this one's going to be a bit of a mess x.x

Player can now attack on clicking mouse 1 using enhanced input shenanigans and animation montages.

The dodge roll animation is totally broken right now so there's a lot of junk relating to  my attempting to retarget animations and fix it.