Root Motion Trial & Error
Went through a tutorial on retargeting animations, but the dodge roll's still jank. Character rolls but the capsule doesn't follow the animation, enabling root motion currently locks the capsule in place until the animation completes; previously it was following the animation but scuffed the rotation, leaving the character facing a little to the left and a little tilted.