Abandoned Dodge Roll Root Motion + Cleaned Up Animation Assets
I've implemented a version of the dodge roll that remains in place as the animation plays, instead of using root motion. I had so many issues that after so much trial and error I've resolved to use code and blueprint to force to the movement I want out of it whilst the animation simply plays. I just couldn't get the player capsule to follow the roll animation without totally messing up the capsule's rotation , so I've resolved to just use this rolling in place animation and set up the movement around it.

I also cleaned up my animation assets after the extensive trial and error ended up producing a lot of unnecessary junk, leaving just the original great sword animations and a few others.