Enemy Blueprint + Consolidated Functionality in Dungeon Character + Migrated Some Animation Assets
Created a new enemy character based on the base dungeon character blueprint. Debating re-parenting it to the dummy blueprint, but haven't done so yet.
Also moved a bunch of functionality into the base dungeon character class, so now all classes can in theory take and deal damage.
Finally moved a bunch of animation assets around just to keep them closer to the character blueprints that are going to be needing them.