Overriding AActor::TakeDamage
Experimented with how to move the AnyDamage event into C++.
Eventually decided to override AActor::TakeDamage() as I was having tons of issues with binding a function to the OnTakeAnyDamage event, despite already achieving a similar result on my health component. Not sure why it refused to work, but I think overriding AActor::TakeDamage() should get me where I want to be anyway.