MVP Reached: AI Always Follows First Target Seen
AI character always follows one player, but otherwise works; they die when they need to, they attack and follow the first player they see, etc. Eventually I'll revisit this to allow for them to change target properly and maybe tether to a location, but that's beyond the scope of the MVP for this project.