Fix Health Component Interactions
Health component is attached in the character's C++ constructor rather than blueprint. Had to duplicate the enemy blueprint to make the change due to unreal errors, which also required the removal of all existing enemies within the level.

I should probably work on a dynamic spawning system at some point. both to avoid the duplication problem in the future and to allow for the game to scale the number of enemies against the number of players.