Enabled Root Motion + Updated Test Dummy
Enabled root motion on the hit reactions and played around with it a bit in the engine.
Also revisited the dummy blueprint to make it easy to test the new animations.

I reparented it to the character class, populated the animations and set it's health to 200 so it'll essentially never die. May be a good idea to create a godmode boolean to make this easier to test in the future.