Flinch Animations Fixed
Flinch animations no longer get characters stuck between states.
This is implemented via a new anim montage that replaces the old hit reaction, and also uses a new anim notify state similar to how melee hit detection is handled.

In bug news the AI is also spinning to face the player when dead, need to check what's causing it and how to stop it.