Fixed Focus After Death (Maybe)
Managed to fix the focus problem I think.
I believe the problem was that sometimes the AI would die and set IsAlive to false after the tree had already started executing that branch, so it would set the focus again one last time before going dormant.
To address this I've placed another IsAlive check on that specific node.

The specific timing here means it's difficult to intentionally recreate this bug, and while I think I've solved it for now I'll have to remain on the lookout in case there's more here I'm missing.