AI Movement Rooted, Turns Still
Both the AI and the player have their feet rooted to the ground when attacking, though the AI has a slight advantage in being able to turn during the animation.
For the moment this is fine, I think I'll be overhauling larger elements of the combat system soon anyway so I'll leave it as is.